generate initial db config and append it to compiled binary
Initial database init and factory reset is slow on STM32 due to the large amount of configurable
parameters. To solve this, a new application is written which uses Database, LESSDB and EmuEEPROM
components but instead of writing to MCU flash, it writes to a binary file on PC. Once this is
done, the resulting file is inserted into binary to flash on MCU on a dedicated page which is used
only as read-only page. Calling factory reset now will erase two used flash pages as it is the case
now, but once this is done, the entire page with initial values will be copied to first used pages,
skipping the initial process entirely.

This also fixes some issues with simulated EEPROM on STM32.