emulated eeprom: set value to 0 if var not found
On AVR, the entire EEPROM is directly addressable and can always be accessed.
When cleared, all values are set to 0. When emulated EEPROM is used, variable
is accessed by going through the entire flash page and finding it by address.
If this value isn't found, reading was considered failed until now. To make things
consistent with AVR, if the variable isn't found, return true and set the value to
0. This also avoids unnecessary failures when performing user-specified data
initialization in database.