system: turn all leds off before calling leds.init()
Done so that it's easier to test led states when system is
initialized. Ideally this wouldn't be needed and it would be
called in leds.init() but then current preset indication
doesn't work on startup - preset change would set the led
state on, then init would be called and the led would be
off again.