refactor display and i2c subsystem
* Implemented I2C registration model similar to the touchscreen model
registration. I2C is now the base class on which components can register.
Display is for now the only I2C component, but support for more
components is planned. This set of changes makes adding such support easy.
* Database/System layout is now changed as well: all display configuration
parameters are now stored in section 0 of I2C block, so this is a breaking
change.
* Various display code cleanups
* Remove PROGMEM qualifier from strings as it causes weird issues on AVR